pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-cli"
version = "0.1.0"
description = "CLI for managing a local markdown knowledge base sourced from external systems."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.32.0",
  "youtube-transcript-api>=1.2.4",
]

[project.optional-dependencies]
crawl4ai = [
  "crawl4ai>=0.6.0",
]

[project.scripts]
know = "knowledge.cli:main"
knowledge = "knowledge.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
